refactor: make SonarCommand platform independent
